cleaner

+++ new.sh
@@ -5,11 +5,16 @@ if [[ $# -eq 0 ]] ; then
 	exit 1
 fi
 
+if [[ ! -e ./templates/$1.cpp ]] ; then
+	echo "doesnt exit"
+	exit 1
+fi
 
 cp ./templates/$1.cpp A.cpp
-rm ./.in/* 2> /dev/null
-rm ./a.out 2> /dev/null
+
+rm -r .in
 
 mkdir .in
+
 gedit A.cpp >/dev/null &
